Refactor fiftystates package name
