+++ fiftystates/src/main/java/com/gluonhq/samples/fiftystates/FiftyStates.java
@@ -24,60 +24,29 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.fiftystates.model;
+package com.gluonhq.samples.fiftystates;
 
-/**
- *
- * Population Density (pop/km2)
- */
-public class Density {
+import com.gluonhq.charm.glisten.application.MobileApplication;
+import javafx.scene.Scene;
+import javafx.scene.image.Image;
+import javafx.stage.Stage;
 
-    public enum DENSITY {
-        D000(0, 10),
-        D010(10, 50),
-        D050(50, 100),
-        D100(100, 250),
-        D250(250, 500),
-        D500(500, 10000);
+public class FiftyStates extends MobileApplication {
 
-        final double ini;
-        final double end;
-
-        private DENSITY(double ini, double end){
-            this.ini = ini;
-            this.end = end;
-        }
-
-        public double getEnd() {
-            return end;
-        }
-
-        public double getIni() {
-            return ini;
-        }
-
+    @Override
+    public void init() {
+        addViewFactory(HOME_VIEW, BasicView::new);
     }
 
-    /**
-     *
-     * @param state
-     * @return DENSITY category for the given US State
-     */
-    public static DENSITY getDensity (USState state) {
-        return getDensity(state.getDensity());
+    @Override
+    public void postInit(Scene scene) {
+        scene.getStylesheets().add(getClass().getResource("style.css").toExternalForm());
+
+        ((Stage) scene.getWindow()).getIcons().add(new Image(FiftyStates.class.getResourceAsStream("/icon.png")));
     }
 
-    /**
-     *
-     * @param density
-     * @return DENSITY category for a given population density
-     */
-    public static DENSITY getDensity (double density) {
-        for (DENSITY d : DENSITY.values()) {
-            if (d.getIni() <= density && density < d.getEnd()) {
-                return d;
-            }
-        }
-        return DENSITY.D000;
+    public static void main(String[] args) {
+        launch(args);
     }
+
 }
