+++ fiftystates/src/main/java/com/gluonhq/samples/fiftystates/model/USState.java
@@ -24,7 +24,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.fiftystates.model;
+package com.gluonhq.samples.fiftystates.model;
 
 /**
  * US State
